Enable new QBluetoothSocket backend for dbus sockets on Bluez 5.46+
The API was introduced earlier than 5.46 already but so far it has not
been tested. For now we use 5.46.

There is another dependency to QLowEnergyController's GATT custom
implementation. Custom GATT only works with the QBluetoothSocket
raw socket implementation. By setting the minimal version for
QBluetoothSocket slightly higher than the 5.42 version for custom
gatt, we ensure that the new dbus socket code is never run together
with the GATT custom stack.

Task-number: QTBUG-68550
Change-Id: I240f7fc8acb116c71e7601df8baf82f61e53c33e
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>