winrt: Make device radio discoverable when listening on socket
Microsoft changed the default behavior so that the device is no longer
radio discoverable, when advertising is started.

By creating the service provider from the Xaml thread we avoid messages,
that it cannot be used in the background by default (even though
this does not seem to have an effect on its functionality).

Change-Id: Ib199adc53ef02b14fcea2884fe4359ad2fe81b8e
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>