qnearfieldmanager_emulator: Use Q_GLOBAL_STATIC for tag activator
Using a Q_GLOBAL_STATIC here fixes an issue in the auto tests for winrt.
If we use a static object that is created before the QCoreapplication has
been created, TagActivator's timer will complain, that it was not created
inside a QThread.
By postponing the creation until first usage we can avoid that problem.

Change-Id: I9beca89930bfca2c71e0fce3d9c000dfcb71f431
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>