Prepare tst_qnearfieldtagtype2 for configurations with builtin_testdata
We have to add relevant test data to TESTDATA and use them from the
created qrc file.

As the emulated TagActivator for configurations with builtin_testdata uses
"nfcdata" as the root it searches for ".nfc" files, the test data has to
be moved into a folder that is called like that. We cannot move the type 2
testdata into the shared nfcdata folder as this would mix up tag types 1
and 2 and thus make the tests fail.

Change-Id: I424dbab5c952408b441f1d67bdcd2fad635f15f4
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>