Fully qualify the parameter in QNearFieldManager::adapterStateChanged
The signal parameter needs to be fully qualfied so that it can be
correctly matched when using the macros.

This can cause a breakage if an application was connecting using the
non fully qualified type, which can be the case if invokeMethod() or
similar is used. However the chances of this are incredibly slim, and
it is worth the risk to ensure normal connections will work correctly.

[ChangeLog][Important Behavior Changes][QtNFC][QNearFieldManager]
adapterStateChanged now has a fully qualified parameter -
QNearFieldManager::AdapterState. This can cause a problem if the
connection/method invocation was done with just AdapterState as
the parameter type, in which case these need to be changed to
use QNearFieldManager::AdapaterState instead.

Change-Id: I7311ff4cb15b67089ceb5917961ee979c9566f28
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>