Bluetooth LE scan - fix a crash (CoreBluetooth)
1. When the central's state changes to powered off, we emit PoweredOffError
and QBluetoothDeviceDiscoveryAgent deletes Obj-C instance - a delegate
for CBCentralManager. But we can still have GDC timer waiting in a queue
and triggering the crash while using a dangling pointer. So we have to properly
cancel the timer.
2. CoreBluetooth under debugger warns about API misuse - calling stopScan,
apparently, is not allowed if CBCentralManager is in a state different
from 'powered on'.

Change-Id: Ib218105735995dc7988751fa04a6c76cab10cba8
Fixes: QTBUG-73140
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>