Fix unit test failure on Android
It fixes the following failure

FAIL! : tst_QBluetoothServiceDiscoveryAgent::tst_invalidBtAddress() Compared values are not the same
     Actual (discoveryAgent->error()) : InvalidBluetoothAdapterError
    Expected (QBluetoothServiceDiscoveryAgent::NoError): NoError

This behavior was caused by the fact that the InvalidBluetoothAdapterError
was already triggered in the QBluetoothServiceDiscoveryAgent ctor whereas
convention for this class states that the error is set when
QBluetoothServiceDiscoveryAgent::start() is called.

The fix detects whether the requested local adapter address matches the
existing local adapter address. If there is no match Invalid adapter error
is thrown.

Task-number: QTBUG-73571
Change-Id: I3216e1609820a66893768b33f0fc695fbad6966a
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>