winrt: Handle device disconnection properly
"ConnectedState" is not the only valid "connected" state for
QLowEnergyController. "disconnected" also has to be emitted, if we were
in DiscoveredState and co.

Additionally every value change callback has to be removed and mDevice
has to be closed in disconnectFromDevice to make Windows disconnect from
the device as there is no specific API to do that manually.

Task-number: QTBUG-74394
Change-Id: Ic025b2c668f887db6c4a631cd8fee26a457bc279
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>