winrt: Stop LE device watcher when discovery is finished
In order to protect ourselves from late callbacks we have to unregister
the LE watcher's callback as soon as we signal, that we are done with
device discovery. We do not expect any additional devices afterwards and
might run into a late callback otherwise, so we should stop the watcher
immediately and not wait until worker destruction.

Change-Id: I85520a4724397aeb5a693d2c8b940a72d68d9663
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>