winrt: Fix notifications in Bluetooth LE
We have to register callbacks for every indicatable characteristics we
encounter during lookup so we have to store our list for the whole
lookup period.

Previously the list of indicatable characteristics was recreated in
every "Completed" handler of Characteristic::GetDescriptorsAsync.
The result was, that at most 1 indicatable characteristic was found.
That behavior is wrong and we have to store the list of indicatable
characteristics through the whole obtainCharList call.

Fixes: QTBUG-74394
Change-Id: Ie93d9bd0184686d2d7cf9b082e85cf89ca5493ce
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>