winrt: More corrections for handling device disconnections
The signal about device disconnections has to be done at the very end of
disconnectFromDevice as a user's application might access an invalid
mDevice otherwise if it tries to reconnect automatically.

The list of device services has to be invalidated when a remote device
disconnects. The same is done in other backends.

Additionally the "valueChanged" callbacks should be unregistered on
device disconnection. Thus we avoid late callbacks which might cause
asserts.

Task-number: QTBUG-74394
Change-Id: I156be8ebefc9d9e5533bc60e7018088641680ba3
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>