IO/CoreBluetooth - refactor the code a bit
... to simplify future code de-duplication. In this patch we:
- For simplicity move all C++ delegate classes into their
  own header. Since we have to hide Objective-C from C++ compiler
  (while compiling *.cpp files), these updated delegates
  have to use void * instead of Objective-C classes.
- Introduce a new RAII classes, that work with Objective-C
  instances but have a header that can be included into the
  *.cpp files, thus making it possible to share *_p.h files with
  all back-ends.

I'm also switching to a new naming convention, which will later
propagate to the pre-existing code - given files are already in
a sub-directory 'osx', having a prefix 'osxbt' in a name is
excessive. Now it's becoming 'bt' (include "osx/btsomething.h").
Later 'osx' dir is to be renamed into 'darwin', which is what
it is these days. Namespace OSXBluetooth is to become DarwinBluetooth.

Task-number: QTBUG-75348
Change-Id: Iebaeab7d0c5e672efebab8218debdec761353633
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>