LE controller/service remove code duplication (Darwin)
Make our private controller class to inherit the shared
base as other platforms do. Rename it into PrivateDarwin.
Remove QLowEnergyController's code duplicate on Darwin.
Get rid of Darwin's copy of QLowEnergyService.
As a bonus, a number of bugs/inconsistencies were fixed
in setting the state and not emitting stateChanged.
Reduced the usage of 'isValid'.

Task-number: QTBUG-75348
Change-Id: I77495870597b61fecae8bca1617590fd9ad2def4
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>