winrt: Avoid threading issues when handling characteristic changes
As callbacks do not necessarily happen on Qt's main thread, it is
possible that we access the service list while it is being changed if we
access it from the callback directly. Doing this can cause application
crashes. Thus we hand back the information about a changed characteristic
to the main thread via signal/slot and handle it from the main thread.

With QLowEnergyControllerPrivateWinRTNew having the Q_OBJECT macro, we
can no longer use forward declares for the GATT classes as moc chokes on
these. Thus an include is used.

Fixes: QTBUG-75907
Change-Id: I063794eecf904921ff55fab76a5bdde3a9aebf44
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>