Fix build with deprecated APIs disabled since 5.13
The DataCompleteness enum has been deprecated since 5.13, but some of
its uses are still compiled for all versions. Make the related code
and docs compile only when the enum is defined.
To keep source compatibility, add another version of serviceUuids()
returning a QList and not taking a DataCompleteness parameter.

Task-number: QTBUG-76491
Change-Id: I20e045230df18e7211ef9982d98ed6e77ec87d16
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>