Modify point when lowenergyscanner remembers found BTLE devices
Technically speaking, nothing really changes in the application. However
devices found, when QBluetoothDeviceInfo::deviceDiscovered() is emitted,
may not have the complete set of data for a given device. In particular,
rssi and manufacturer data changes happen later.

Later changes to the example may want to display these transient types
of information at which point this change becomes essential. It makes
the app simpler to adopt for testing purposes too.

Change-Id: I119b6e80a9a92b16de2d3e6792132ceb06b577a4
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: André de la Rocha <andre.rocha@qt.io>