Bluetooth: change prefix in Obj-C class names
As a follow-up of the recent file/namespace re-naming in QtBluetooth.

Previously, the prefix was 'OSXBT'. This prefix is needed because:
- Obj-C classes cannot be defined inside a C++ namespace, thus
- A very generic name like 'DeviceInquiry' is potentially problematic.
- We use QT_MANGLE_NAMESPACE, but for non-namespaced build it does not help.

To make these names somewhat unique, I've introduced 'OSXBT' prefix with
'OSX' for a platform and 'BT' for 'Bluetooth'. 'OSX' part of the prefix
is now outdated and I'm replacing it with 'Darwin'. Also, if a resulting
name is still ambiguous I add 'Classic' to the name ('LE' was already
in place where needed).

Change-Id: I4cb6069c96f4159dd6cc0f94823454ea6c82b1ef
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>