Add descriptor even if the read fails during service discovery
leDescriptorRead() serves two pruposes. Firstly, it informs Qt of the
value/result of the read operation. During service discovery it is also
used to inform Qt that a descriptor was found. This info is used
to build up the internal data structure representing the service, its
characteristics and its descriptors.

If we have a non-readable descriptor the read failed (as expected)
and Qt was never informed about the existence of the descriptor. During
service discovery the primary purpose is to inform about the existence
of a descriptor though.

This patch ensures the notification comes through and we accept the fact
that the passed value is not known.

Fixes: QTBUG-78201
Change-Id: Id73a27c90905cb769f1cd168b48299d580587594
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 7390de230e3c86049824bef756e4af623a547d61)
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>