bluez: Fix out of sync property cache
Since commit 2c6dcc643, a map containing dbus property values is kept
per device to be able to make QBluetoothDeviceInfo instances without
having to make dbus calls. However, this map could get out of sync. The
reason was that the code started to listen to PropertyChanged signals
after it received the initial set of properties. This meant that there
was a short amount of time where bluez could send out a signal with
changed properties and the DiscoveryAgent would not receive it. In order
to keep the map in sync, a PropertyChanged listener is now installed
when scanning is started, picking up all changes.

Change-Id: Iae5e219b187e94bf280a6dd76ccde49cafd1d9c5
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>