Fix up broken PropertiesChanged connections
Change b7b979d1735764930cc6f736cfa8e22b71547ea8 introduced a slightly
modified version of the OrgFreedesktopDBusPropertiesInterface::PropertiesChanged
signal. This signal was taken into use in one case but all other
instances of this signal were not modified and broken.

This patch modifies the remaining cases and converts the connect statements
to function pointer syntax. This forces the compiler to check that signal
and slot parameters match.

Change-Id: I41b45f7e2f7b66ff29f321acc38ff2ce1c81864e
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Thiemo van Engelen <tvanengelen@victronenergy.com>