Fix a race condition between a write and an incoming change
The Android API that is available to write to a characteristic requires
the user to first set the new value of the characteristic and then tell
Android to write the value to the device. If the phone processes an
incoming change for this same characteristic in between the setting of
the value and the actual write, it could happen that the characteristic
gets the incoming value and that incoming value is then sent to the
connected device. Something similar could occur for the incoming change
as the received value is not passed in the callback but it is set on the
characteristic and the user is expected to read the received value from
the characteristic.
To prevent these problems, a new connectGatt function that takes a
Handler as extra argument was added to SDK v26. The callbacks are then
always executed via this handler and thus on the associated handler
thread. If the user also performs all writes on this thread, the
described race condition can no longer occur.
This could have solved the problems, but Android 8.0 contained a race
condition in the callback part because it set the value of the
characteristic before passing the callback to the handler, still
allowing the same race condition for incoming changed. This was fixed in
Android 8.1.

This commit causes the code to use a Handler and the new connectGatt
function if it is running on SDK >= 27, which is associated with Android
8.1.
For older SDK versions, a different solution is implemented. In this
case, a temporary BluetoothGattCharacteristic is instantiated that can
hold the value to be written. This way, the write can no longer
interfere with changes that are being received. To instantiate the
BluetoothGattCharacteristic, a private constructor is invoked using
reflection. This private constructor is needed as only this constructor
allows to create an instance that has the required information for
Android to be able to write to the charatceristic (such as the
associated service and instanceId). This is also the reason why this
solution cannot be used on newer SDK's as this constructor is
blacklisted and a warning box will be shown when reflection is used to
get a reference to it. Because a temporary BluetoothGattCharacteristic
is instantiated that holds the written value, it is necessary to store
the value that was written and pass that in the callback.

More information on these Android issues and solution direction can be
found here:
https://stackoverflow.com/questions/38922639/how-could-i-achieve-maximum-thread-safety-with-a-read-write-ble-gatt-characteris
https://medium.com/@martijn.van.welie/making-android-ble-work-part-3-117d3a8aee23

Change-Id: I5a9eda501a20933e37f758a3114bf71ec3e7cfd4
Reviewed-by: Thiemo van Engelen <tvanengelen@victronenergy.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>