Add location-turned-on check before starting device discovery
The location permission and a running location service have
been a prerequisite for a successful LE scan since Android v23.
While the permission has always been checked before doing a bluetooth scan, the
state of the location service was not checked. LE discovery
without location turned on does not error out. It just reports zero
discoveries.

Starting with Android SDK v29, a classic Bluetooth discovery requires
a running Location service too. Due to BluetoothAdapater.startDiscovery()
returning an error code when location is turned off, it was time to
check the location service before triggering the discovery
to simplify the error handling for the API user.

This patch ensures that location is turned on before any type of device
discovery is started and the API reports an error reflecting this problem.

Task-number: QTBUG-81875
Change-Id: Id51fd502ae24cbadbc704451fdf49d999224c16f
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>