Android: Ensure that LE errors are forwarded from JNI to Qt
The code to detect the error existed on the Java/JNI side. The code to
handle and report the error was available on the Qt side. Unfortunately
the essential signal and slot connection was never made though. This lead
to

[ChangeLog][QtBluetooth][Android] Fixed the missing
QLowEnergyService::error() signal emission if the error was detected
on the Java side. This probably cut 50% of all errors out.

Change-Id: I6ee4d7605a1d1e61b920da56d7f67373fb7f5be5
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>