Remove function sendCommands(...) from QNearFieldTarget
The result of the function is different from that specified in the
documentation. Its only the result of the last command and not a list.

Also the purpose of the function is questionable. An additional
command should only be send, when the last command was successful.
This would be possible by providing a list of pairs (command,
expectedResponse). But this is dangerous/complex because the response
is not always fully known.

[ChangeLog][QtNFC][QNearFieldTarget] Removed function sendCommands(...)

Task-number: QTBUG-81824
Change-Id: I670f59ad74aacfb5c643b4c8fffb792dc93bfb06
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>