Remove functions keepConnection() and setKeepConnection(...)
This functions were introduced in d977db4 to stay backward
compatible. With Qt 6 we can change this behavior and keep
the connection by default.

The normal usecase is to delete the QNearFieldTarget after
the target appeared and the communication is done. So we
can disconnect the target in the destructor.

In the rare cases where the communication should be continued
after a different instance communicated with the target,
disconnect can be called by the user to allow the other
tag to communicate.

[ChangeLog][QtNFC][QNearFieldTarget] Removed functions
keepConnection() and setKeepConntection(...)

Task-number: QTBUG-81824
Change-Id: I32d23d07f3df9be11127d4dc6c7432627877b22e
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>