Remove QNearFieldShareManager/QNearFieldShareTarget
The API Was introduced in dd75b1f77 with an implementation for
Blackberry only. Support for Blackberry was removed in 3e5a32e9
keeping the interface only for backward compatibility.
There is no future for Android because the beam methods are
deprecated since API 29 with the following comment:
"File sharing can work using other technology like Bluetooth"

[ChangeLog][QtNFC] Removed QNearFieldShareManager/QNearFieldShareTarget

Task-number: QTBUG-81824
Change-Id: Ib0875c33072494637af42e868939a8c301324aba
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>