Only scan for NFC tags with the given access method
On iOS it is only possible to search for one type of tags at the time.
See https://developer.apple.com/documentation/corenfc/nfcreadersession

On Android the access mode is controlled by the Intent filter the user
needs to set. So we check the target before targetDetected is emitted.
In the future we will add the readerMode, so we can also avoid the
scanning for wrong tags.
https://developer.android.com/reference/android/nfc/NfcAdapter#enableReaderMode

Task-number: QTBUG-81824
Change-Id: I5aa2580ada619363560cdad354a3a0be4998ea96
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>