Update commercial license headers
Updated header.COMM to the files in tqtc-qtconnectivity.
Examples, tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in test.

Task-number: QTQAINFRA-4159
Change-Id: Icf450df25282ffadf5bcd0950753d17ba6ab57e3
Reviewed-by: Antti Kokko <antti.kokko@qt.io>