Add server role to BlueZ service registration
In service registration, a profile is created. This profile had
no clien/server flag. If two profiles with the same UUID are
created (one for server side, one for client side) this led to
the wrong profile being used in connection setup.
This patch marks the profile which is created in registerService()
as a server profile which prevents it from being used by BlueZ
in a client context.

Change-Id: I95b22657a7c7e45b334476a8513d79920ba9cc96
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit e74cb54a4cbc8aba188ea48ae3effcf10432df11)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>