Correct error handling in QLowEnergyControllerPrivateBluezDBUS
When a DBUS interface was removed for a device, error handling
just assumed it was the org.bluez.Device1 interface, without checking.
This lead to errors on disconnection if the device has a battery
service: The code sends a disconnection request, then an
InterfaceRemoved event for org.bluez.Battery1 occurs. This was
interpreted as a removal of org.bluez.Device1 which in turn
led to the controller going into UnknownRemoteDeviceError error state.
This patch adds a check whether it is really org.bluez.Device1
which is removed.

Change-Id: I449b29cb9528cda23ce972f36f716a8774f01fe5
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit af5f801c6b8de0bf0d21798bddef54cb14759ecd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>