Fix unit test flakyness by only detecting number of devices once
tst_QBluetoothLocalDevice was flaky. It turns out it queried
the number of available bluetooth adapters in each test. When
no bluetooth adapter is present, this call first tries to
contact Bluez5 and times out after about 20 seconds, then it
tries to contact Bluez4 and times out after about 20 seconds.
This patch fixes the problem by only asking for the number of
bluetooth adapters once and then reusing the result.
A better fix will be available in the future, when Bluez4
support will be dropped.

Change-Id: I2c478479f754bcc8b08e83560b5a78c6060e2329
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>