Improve log message when setting negative BTLE discovery timeout
Setting the BTLE discovery timeout fails if a timeout is not
supported or if the user tries to set a negative timeout. The
log message did not tell which of those two cases happened.
This patch improves the log message to have different messages for
those two cases.

Task-number: QTBUG-90369
Change-Id: I6b87259557e71a6a18f0604d299fda2788c6bd32
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>