Fix tst_qbluetoothdevicediscoveryagent after BlueZ 4 removal
Change in initialization in 87f6d3603bdcadb96dcce98e8eb46a25e5184767
broke tst_qbluetoothdevicediscoveryagent. The test was looking for
the string "Bluez 5" in the debug output, which is not produced any
more. This patch removes the logic to distinguish between BlueZ 4
and BlueZ 5 from the test, always assuming BlueZ 5.

Change-Id: I8f6cba66ac89f03493a765f4f32b868970e9b666
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>