Reduce timing issues during manual QBluetoothLocalDevice pairing tests
The manual part of the test requires the user to accept pairing requests
on both devices. Typically this requires the acceptance of a system popup.
The user action requires time and the given timeouts were not sufficient.

Furthermore the use of QTest::qWait() is inefficient because it
waits exactly for the requested amount of time without taking into
consideration that the process may have finished (as highlighted by the
relevant signals). That's the task of QTRY_VERIFY_WITH_TIMEOUT.

30s is the wait time for waits related to manual input and 5s is the
timeout for automatic sections. The latter value is based on the fact
that the previous QTRY_VERIFY uses those 5s waiting time too.

In summary, the test increases timeouts but at the same time blocks the
entire test for a much shorter time frame (and therefore requiring
much less time to execute in total).

Change-Id: If2b21af4949f34c4f15e7794e21c002b24605e91
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>