Set android nfc transceive timeout
As it stands, each device uses its own timeout for the transceive command.
While most devices seem to use 618 ms, there are some devices which choose
a timeout of something like 309 ms.
So lets set a common timeout, so the behavior stays consistent between
devices. Use 2000 ms as that should be enough for most applications.

Change-Id: I4fae8f284b9df6c78f7117e966dd9fed2f48fd9d
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>