Add manual BT LE test
This patch adds a unit test which tests various BT LE features.
The test assumes to find a device running the "bluetoothtestdevice"
BT LE server from this same repository.

Finding/matching the server device is based on the BT LE name
which can be provided by an environment variable or by hardcoding
it in the testcase code.

Pick-to: 6.2
Change-Id: I1c6a1bad168066d274c95e6f2cafb3ef5f9b48c1
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>