CoreBluetooth: add a workaround to enable using scan options
By default, the value for CBCentralManagerScanOptionAllowDuplicatesKey
is 'NO' (but either passing it explicitly or passing nil as 'otpions').
According to a report, this disables a proper device discovery (contrary
to what is documented by Apple). Since, according to the Apple's docs,
"Disabling this filtering can have an adverse effect on battery life;
use it only if necessary", we don't set it by default and instead
provide a workaround using environment variable to set the option,
if needed. The name of this variable is:
"QT_BLUETOOTH_SCAN_ENABLE_DUPLICATES".

[ChangeLog][QtConnecivity][Bluetooth][Darwin] Enable setting a scan option (that allows duplicates)

Pick-to: 5.15
Change-Id: I08a5a089020c433a87bcb622d08aa2cd260283b9
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>