Add convenience API to access CCCD
Notification or indication for a characteristic can
be enabled/disabled by writing special bit patterns
into the CCCD, the client characteristic configuration
descriptor.
Until now, it was necessary to obtain this descriptor
using its UUID and then knowing the bit patterns
to write into it.
This patch adds a method named "cccd" to obtain
the cccd and adds named constants for
enabling and disabling indication and notification.
Thus, the developer does not have to look them up.

Fixes: QTBUG-92986
Task-number: QTBUG-75341
Change-Id: I8c510d5ef66909bcf6e3885076bb3629767a9064
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>