QNdefFilter: add match() method
The method itself behaves differently depending on the orderedMatch
parameter. If orderedMatch is false, we can just group all the
filter records with the same typeNameFormat and type, and then
calculate the actual number of occurrences for each filter record.

If orderedMatching is enabled, we need to take positions into
account. This also means that we need to merge consecutive records
with the same typeNameFormat and type, because otherwise we would
not be able to check, which filter record to use for the match.

[ChangeLog][QtNFC][QNdefFilter] Add QNdefFilter::match() method that
can be used to check, if the message matches the filter.

Task-number: QTBUG-93854
Change-Id: Ib219fc0e520af4a50c9027035a529119714d6d18
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>