BTLE Android advertising data size limit: improve documentation
BTLE advertising just fails on Android if advertising data is too large.
This is unfortunate, as the size of the advertising data cannot
be predicted without knowledge of the advertising data format.
This patch introduces at least a warning in the documentation that
this problem exists and adds an error message which tells
the developer what to do.

Task-number: QTBUG-93595
Change-Id: I21d69fd314aa9165c223cbdb2394dca6826afdd6
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>