AnnotatedUrl: refactor UI
This patch refactors the UI layout, so that it handles screen rotation
correctly. The *.ui files are removed. The layout is created directly
from the source code instead.
The application now has two different labels to show the icon.
Depending on the actual screen orientation, one of them shows the icon,
and the other is hidden.
The icon is properly scaled before it's shown, so that the layout does
not expand beyond the visible screen rectangle.
Also this patch introduces a code that handles screen orientation
change, and the screen change (in case the application is dragged to
the other monitor). The latter is not really applicable to mobile
devices though.

Task-number: QTBUG-94033
Change-Id: I0d7d781bd2b97e16f7fba530eeffd407c7eaff61
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 6d4d0ca9ea79ae64ab1bea33e9b99a76430ffc3c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>