Improve detection of enabled location service on SDK<28
Bluetooth seems to be non-functional when location services are
disabled. This is not documented in Android's documentations,
but several bug entries exist about this which are closed
as "works as intended". See the linked bug entry for details.
Our code thus tries to detect whether location services are enabled.
This patch improves the detection of enabled location services.
While the previous code did only check for GPS and Network
location providers, the new code should detect any enabled
location provider.

Fixes: QTBUG-90760
Pick-to: 5.15 6.2
Change-Id: I0c670296d0af62161dbc3fca40889996fa91f0ca
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>