Rename streamingoperator to adhere to camelCase
operator<< was changed to be a hidden friend in QtBluetooth.
Two properties of operator<< were desired:
First, it should be a hidden friend. This implies that
it is defined inline.
Second, its implementation should not part of the public interface.
To achieve both, an indirection and the static method
streamingoperator() were introduced.
This patch changes its name to streamingOperator() to
adhere to Qt's camelCase coding style.

Pick-to: 6.2
Change-Id: I4e01a81eac8261c5198cdb832c1345efbf647419
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>