Android: Call handleTargetLost directly on NFC TagLostException
If in a transceive() a TagLostException occurs, subsequent calls to
transceive() may lead to an app crash with "JNI DETECTED ERROR IN
APPLICATION: GetMethodID received NULL jclass" on Android 11 if the
timer did not clean up first.

This fix calls handleTargetLost directly. This could lead to a race
condition in which the targetLost signal is send twice, which should be
handled correctly by the Qt code.

Pick-to: 5.15 6.2
Change-Id: I886e0bd436272ee393a92b0193fd4d40ed4e965e
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Lars Schmertmann <lars.schmertmann@governikus.de>